DownsampleScale now proper uint in shader
No need for it to be float. This communicates more clearly the intend and I'm pretty sure pref is not much worse (if not faster).

+++ Resources/Shaders/Merge.frag
@@ -18,7 +18,7 @@ uniform vec3 inverseSecondaryLightDirection;
 
 uniform float zNear;
 uniform float zFar;
-uniform float downsampleScale;
+uniform uint downsampleScale;
 
 float calcLinearDepth(float bufferDepth) {
   float normalizedDepth = bufferDepth*2.0-1.0;
